+++ CHANGES.md
@@ -8,7 +8,9 @@ E-Maj - Change log
   * Minor code changes.
 
 ###Bug fixes:###
-
+  * Fix the records order for emaj_snap_group(), emaj_dump_changes_group()
+    and emaj_gen_sql_dump_changes_group() outputs for some tables having a
+    multi-column PK.
 
 
 4.7.0 (2025-Aug-31)